requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
attrs==19.3.0             # via hypothesis, jsonschema, pytest, schemathesis
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via -r requirements/dev.in, flask, mkdocs, nltk, pip-tools, schemathesis
flask==1.1.2              # via -r requirements/dev.in
future==0.18.2            # via lunr
graphql-core==3.1.2       # via hypothesis-graphql
hypothesis-graphql==0.3.2  # via schemathesis
hypothesis-jsonschema==0.18.0  # via schemathesis
hypothesis==5.38.0        # via hypothesis-graphql, hypothesis-jsonschema, schemathesis
idna==2.10                # via requests, yarl
iniconfig==1.1.1          # via pytest
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask, mkdocs, mkdocs-macros-plugin
joblib==0.17.0            # via nltk
jsonschema==3.2.0         # via hypothesis-jsonschema, schemathesis
junit-xml==1.9            # via schemathesis
livereload==2.6.3         # via mkdocs
lunr[languages]==0.5.8    # via mkdocs
markdown==3.3.3           # via mkdocs, mkdocs-material, pymdown-extensions
markupsafe==1.1.1         # via jinja2
mkdocs-macros-plugin==0.4.18  # via -r requirements/docs.in
mkdocs-material-extensions==1.0.1  # via mkdocs-material
mkdocs-material==6.1.0    # via -r requirements/docs.in, mkdocs-macros-plugin, mkdocs-material-extensions
mkdocs==1.1.2             # via -r requirements/docs.in, mkdocs-macros-plugin, mkdocs-material
multidict==5.0.0          # via yarl
nltk==3.5                 # via lunr
packaging==20.4           # via pytest
pip-tools==5.3.1          # via -r requirements/dev.in
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pygments==2.7.2           # via -r requirements/docs.in, mkdocs-material
pymdown-extensions==8.0.1  # via -r requirements/docs.in, mkdocs-material
pyparsing==2.4.7          # via packaging
pyrsistent==0.17.3        # via jsonschema
pytest-subtests==0.3.2    # via schemathesis
pytest==6.1.1             # via -r requirements/tests.in, pytest-subtests, schemathesis
python-dateutil==2.8.1    # via mkdocs-macros-plugin
pyyaml==5.3.1             # via -r requirements/dev.in, mkdocs, mkdocs-macros-plugin, schemathesis
regex==2020.10.23         # via nltk
requests==2.24.0          # via schemathesis
schemathesis==2.6.0       # via -r requirements/tests.in
six==1.15.0               # via jsonschema, junit-xml, livereload, lunr, packaging, pip-tools, python-dateutil
sortedcontainers==2.2.2   # via hypothesis
starlette==0.13.8         # via schemathesis
termcolor==1.1.0          # via mkdocs-macros-plugin
toml==0.10.1              # via pytest
tornado==6.0.4            # via livereload, mkdocs
tqdm==4.51.0              # via nltk
urllib3==1.25.11          # via requests
werkzeug==1.0.1           # via flask, schemathesis
yarl==1.6.2               # via schemathesis

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
